pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "psycopg-binary>=3.2.10",
    "pydantic>=2.11.9",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.35.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    "gspread>=6.1.4",
    "google-auth>=2.35.0",
]



[tool.aerich]

tortoise_orm = "core.db.TORTOISE_ORM"

location = "./migrations"

src_folder = "./."

